Implement the dispatch of payment request events to service workers.
This patch implements ServiceWorkerManager::PropagatePaymentRequestEvent which
is supposed to be called through nsIPaymentRequestService on the main thread in
the chrome process when PaymentRequest::Show() is called and the payment
handler is selected.

First, this message is passed to ServiceWorkerManagerService (SWMS) on the
background thread in chrome process through PServiceWorkerManager. Second,
registered PServiceWorkerManager child actors on content processes will then
receive this message from SWMS. Third, service worker managers on those content
processes will use the WorkerPrivate with given origin and scope to dispatch
the PaymentRequestEvent to target service workers.

Note that in this patch, the event is dispatched to all content processes. A
further improvement could be made by picking just one content process to
dispatch the event after Bug 1368625 is done.